Revert "Merge branch 'main' into group-and-game"
This reverts commit 8c768a648f1dc6c07243e47f700e43a2a20a6021, reversing
changes made to 6cd2ed4dc6322b135f97e7ee1f0b2d8dc2a326a6.

+++ src/main/java/io/dot/jyp/server/application/dto/GroupCreateResponse.java
@@ -8,11 +8,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class GroupCreateResponse {
-    private Long id;
+    private int id;
     private String name;
 
     public static GroupCreateResponse of(
-            Long id,
+            int id,
             String name
     ) {
         return new GroupCreateResponse(
